MFC r349019
bhyve: move common code to net_utils.c

Both virtio_net and e82545 network frontends have code to validate and
generate MAC addresses. These functionalities are replicated in the two
files, so we move them in a separate compilation unit.

Reviewed by:    rgrimes, bryanv, imp, kevans
Differential Revision:  https://reviews.freebsd.org/D20626
